Back_LLM/pyproject.toml
[project]
name = "back-lmm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.0",
    "openai>=2.7.1",
    "uvicorn>=0.32.1",
    "pydantic>=2.10.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
]
